added profiler and refactored code base
